CMake: Cleanup of cmake files
intrinsics:
Changed to use _tzcnt for ctz (count trailing zeros), which should be available on every x64 processor.

ext:
Moved folder out from src.

molecule:
removed free function from api and moved into common free function for molecule structure.
changed backbone to be independent of chain.
added instance data.

pdb:
Added assemblies which contain instance information extracted from REMARK 350 BIOMT

str:
renamed copy_str to str_copy.

gl:
changed to use char for opaque byte blobs.
some changes propagated from molecule backbone structure changes.

os:
Added path_exists and path_is_directory.

simd:
Added movemask

str:
Ranamed functions to be more coherent with str_ prefix

vec_math:
Changed unions into structures of unions to be coherent for pre-declarations.

util:
Revamped postprocessing to include flags to determine what data to compute.
This is used primarily to have a subset which suits Coarse Grained datasets.

edr:
Started work on gromacs edr file reading functionality to read energy parameters.

gfx:
Started work on new graphics component which should be an agnostic front-end and support modern OpenGL 4.6 features with indirect drawcall dispatch with occlusion culling.

script:
Added additional check inside element_str since molecule can now miss the field element in case of Coarse Grained data.

+++ src/core/md_common.h
@@ -66,7 +66,7 @@
 #endif // NDEBUG
 
 #ifndef ARRAY_SIZE
-#define ARRAY_SIZE(x) ((int64_t)(sizeof(x)/sizeof(x[0])))
+#define ARRAY_SIZE(x) (sizeof(x)/sizeof(x[0]))
 #endif
 
 #ifndef IS_ALIGNED
